Actualización 12 de enero de 2022

+++ app/Http/Controllers/PermissionController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\CreatePermissionRequest;
 use App\Http\Requests\EditPermissionRequest;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
 
 class PermissionController extends Controller
@@ -16,6 +17,7 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('permission_index'), 403);
         $permissions = Permission::all();
 
         return view('back.permissions.index', compact('permissions'));
@@ -28,6 +30,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('permission_create'), 403);
         return view('back.permissions.create');
     }
 
@@ -51,6 +54,7 @@ class PermissionController extends Controller
      */
     public function show(Permission $permission)
     {
+        abort_if(Gate::denies('permission_show'), 403);
         return view('back.permissions.show', compact('permission'));
     }
 
@@ -62,6 +66,7 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
+        abort_if(Gate::denies('permission_edit'), 403);
         return view('back.permissions.edit', compact('permission'));
     }
 
@@ -86,6 +91,7 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
+        abort_if(Gate::denies('permission_destroy'), 403);
         $permission->delete();
         return redirect()->route('permissions.index')->with('success', 'Permiso eliminado correctamente.');
     }
